chatbot/requirements.txt
fastapi
pydantic
openai
requests
Pillow
uvicorn
python-dotenv

# pip install -r requirements.txt